Separate out router logic from HTTP API logic into separate files.
